+++ doc/source/changes.rst
@@ -15,8 +15,8 @@ New in 0.10
 * Tools for generating colors for categories (see :ref:`ref-plotting`).
 * Plots now all largely respect matplotlib rc-parameters (see
   `Customizing Matplotlib <http://matplotlib.org/users/customizing.html>`_).
-* Fixed a memory leak in the :mod:`testnd` module that could slightly affect
-  permutation based p-values.
+* Fixed an issue in the :mod:`testnd` module that could affect permutation based
+  p-values when multiprocessing was used.
 
 
 New in 0.9